requirements.txt
# Crawling
requests>=2.33.0
beautifulsoup4>=4.14.0
pdfplumber>=0.11.0

# Vector store & embeddings
chromadb>=1.5.0
ollama>=0.6.0

# RAG pipeline (used in rag_chatbot/)
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.3.0
openai>=1.0.0

# Web interface
gradio>=4.0.0

# Reranking (local cross-encoder, no API key required)
flashrank>=0.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
